requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
numpy==1.24.3
scipy==1.11.1
torch==1.13.1+cu117
matplotlib==3.7.1
dill==0.3.6
docopt==0.6.2
einops==0.6.1
backpack_for_pytorch==1.5.2
laplace_torch==0.1a2
